added event bus config and implemented mpsc

+++ includes/topic.hpp
@@ -5,7 +5,7 @@
 namespace eventbus {
     class Topic {
     public:
-        explicit Topic(std::string name, const int partition_count):
+        explicit Topic(std::string name, const size_t partition_count):
         name_(std::move(name)),
         partition_count_(partition_count){}
 
@@ -14,13 +14,13 @@ namespace eventbus {
             return name_;
         }
 
-        [[nodiscard]] int partition_count() const {
+        [[nodiscard]] size_t partition_count() const {
             return partition_count_;
         }
 
     private:
         std::string name_;
-        int partition_count_;
+        size_t partition_count_;
     };
 }
